Reorganize code
I added comments and stored data in different more understandable
folders.
