+++ AgileNLP/nbproject/project.properties
@@ -55,7 +55,7 @@ javadoc.splitindex=true
 javadoc.use=true
 javadoc.version=false
 javadoc.windowtitle=
-main.class=agilenlp.TestHarness.TestHarness
+main.class=TestHarness.TestHarness
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
