+++ AgileNLP/src/FileReader/Movie.java
@@ -1,5 +1,5 @@
 
-package agilenlp;
+package FileReader;
 
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
