+++ AgileNLP/src/FileReader/MovieCharacter.java
@@ -9,7 +9,7 @@ DATA EXAMPLE
 */
 
 
-package agilenlp;
+package FileReader;
 
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
