+++ AgileNLP/src/TestHarness/OneWordParseTest.java
@@ -3,9 +3,8 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package agilenlp.TestHarness;
+package TestHarness;
 
-import agilenlp.*;
 import java.util.ArrayList;
 import java.util.Hashtable;
 import java.util.Scanner;
